# docs/swagger/swagger.yaml
basePath: /v1
consumes:
    - application/json
definitions:
...
    User:
        properties:
            created_at:
                description: The timestamp of when the user was created.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: CreatedAt
            deleted_at:
                description: The timestamp of when the user was deleted.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: DeletedAt
            email:
                description: The email of the user.
                example: '"jhon.doe@example.com"'
                type: string
                x-go-name: Email
            id:
                description: The ID of the user.
                example: 1
                format: uint64
                type: integer
                x-go-name: ID
            name:
                description: The name of the user.
                example: '"John Doe"'
                type: string
                x-go-name: Name
            updated_at:
                description: The timestamp of when the user was last updated.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: UpdatedAt
            uuid:
                description: The UUID of the user.
                example: '"f3b3b3b3-3b3b-3b3b-3b3b-3b3b3b3b3b3b"'
                type: string
                x-go-name: UUID
        required:
            - id
            - uuid
            - email
            - name
        title: User represents a user in the system.
        type: object
        x-go-package: github.com/ryo-arima/locky/pkg/entity/response
    UserRequest:
        properties:
            created_at:
                description: The timestamp of when the user was created.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: CreatedAt
            deleted_at:
                description: The timestamp of when the user was deleted.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: DeletedAt
            email:
                description: The email of the user.
                example: '"jhon.doe@example.com"'
                type: string
                x-go-name: Email
            id:
                description: The ID of the user.
                example: 1
                format: uint64
                type: integer
                x-go-name: ID
            name:
                description: The name of the user.
                example: '"John Doe"'
                type: string
                x-go-name: Name
            password:
                description: The password of the user.
                example: '"password"'
                type: string
                x-go-name: Password
            updated_at:
                description: The timestamp of when the user was last updated.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: UpdatedAt
            uuid:
                description: The UUID of the user.
                example: '"f3b3b3b3-3b3b-3b3b-3b3b-3b3b3b3b3b3b"'
                type: string
                x-go-name: UUID
        required:
            - name
            - email
            - password
        title: UserRequest represents the request body for user-related operations.
        type: object
        x-go-package: github.com/ryo-arima/locky/pkg/entity/request
    UserResponse:
        properties:
            code:
                description: The response code.
                example: '"SUCCESS"'
                type: string
                x-go-name: Code
            message:
                description: The response message.
                example: '"Users retrieved successfully"'
                type: string
                x-go-name: Message
            users:
                description: The list of users.
                items:
                    $ref: '#/definitions/User'
                type: array
                x-go-name: Users
        required:
            - code
            - message
            - users
        title: UserResponse represents the response body for user-related operations.
        type: object
        x-go-package: github.com/ryo-arima/locky/pkg/entity/response
host: localhost:8080
info:
    description: '# This is a sample Locky API'
    title: Locky API.
    version: 1.0.0
paths:
...
    /private/users:
        get:
            description: Get a list of all users in the system.
            operationId: getUsersPrivate
            responses:
                "200":
                    description: A list of users.
                    schema:
                        $ref: '#/definitions/UserResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/UserResponse'
            summary: Get a list of users.
            tags:
                - users
        post:
            description: Create a new user with the provided information.
            operationId: createUserPrivate
            parameters:
                - description: The user to create.
                  in: body
                  name: user
                  required: true
                  schema:
                    $ref: '#/definitions/UserRequest'
            responses:
                "200":
                    description: The created user.
                    schema:
                        $ref: '#/definitions/UserResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/UserResponse'
            summary: Create a new user.
            tags:
                - users
    /private/users/{id}:
        delete:
            description: Delete a user by ID.
            operationId: deleteUserPrivate
            parameters:
                - description: The ID of the user to delete.
                  in: path
                  name: id
                  required: true
                  type: integer
            responses:
                "200":
                    description: The deleted user.
                    schema:
                        $ref: '#/definitions/UserResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/UserResponse'
            summary: Delete a user.
            tags:
                - users
        put:
            description: Update a user with the provided information.
            operationId: updateUserPrivate
            parameters:
                - description: The ID of the user to update.
                  in: path
                  name: id
                  required: true
                  type: integer
                - description: The user to update.
                  in: body
                  name: user
                  required: true
                  schema:
                    $ref: '#/definitions/UserRequest'
            responses:
                "200":
                    description: The updated user.
                    schema:
                        $ref: '#/definitions/UserResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/UserResponse'
            summary: Update a user.
            tags:
                - users
...
    /public/users:
        get:
            description: Get a list of all users in the system.
            operationId: getUsers
            responses:
                "200":
                    description: A list of users.
                    schema:
                        $ref: '#/definitions/UserResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/UserResponse'
            summary: Get a list of users.
            tags:
                - users
        post:
            description: Create a new user with the provided information.
            operationId: createUser
            parameters:
                - description: The user to create.
                  in: body
                  name: user
                  required: true
                  schema:
                    $ref: '#/definitions/UserRequest'
            responses:
                "200":
                    description: The created user.
                    schema:
                        $ref: '#/definitions/UserResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/UserResponse'
                "500":
                    description: Internal server error.
                    schema:
                        $ref: '#/definitions/UserResponse'
            summary: Create a new user.
            tags:
                - users
produces:
    - application/json
schemes:
    - http
swagger: "2.0"